content/requirements.txt
fastapi
psycopg2-binary
sentence-transformers
google-generativeai
pgvector
python-dotenv
pydantic[email]
uvicorn[standard]
python-jose